+++ app/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import React, { useState, useRef, useEffect } from 'react'
+import React, { useState } from 'react'
 import {
   useSurveyState,
   useScheduleState,
@@ -8,11 +8,7 @@ import {
   useChatState,
   useCalendarUrls,
 } from '@/lib/persistence-hooks'
-import {
-  icalEventsToScheduleItems,
-  type ICalEvent,
-} from '@/lib/ical-parser'
-import { WSU_SEMESTER } from '@/lib/constants'
+import { useBackgroundIcsSync } from '@/lib/use-background-ics-sync'
 import type { ScheduleItem, ScheduleItems } from '@/lib/schemas'
 import SurveyView from '@/ui/views/SurveyView'
 import ChatView from '@/ui/views/ChatView'
@@ -63,72 +59,7 @@ export default function ScheduleApp() {
     currentDate instanceof Date ? currentDate : new Date(currentDate)
 
   // Daily ICS calendar refresh - syncs every 24 hours when ICS URLs are configured
-  const scheduleItemsRef = useRef(scheduleItems)
-  const nextTaskIdRef = useRef(nextTaskId)
-  const icsUrlsRef = useRef(icsUrls)
-  scheduleItemsRef.current = scheduleItems
-  nextTaskIdRef.current = nextTaskId
-  icsUrlsRef.current = icsUrls
-
-  useEffect(() => {
-    if (icsUrls.length === 0) return
-
-    const runBackgroundSync = async () => {
-      const urls = icsUrlsRef.current
-      if (urls.length === 0) return
-      try {
-        let currentSchedule = { ...scheduleItemsRef.current }
-        let currentNextId = nextTaskIdRef.current
-        for (const url of urls) {
-          const res = await fetch('/api/fetch-ics', {
-            method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify({ url }),
-          })
-          const data = await res.json()
-          if (!data.success) return
-          const events = (data.events || []).map(
-            (e: {
-              uid: string
-              title: string
-              start: string
-              end: string
-              isAllDay: boolean
-              location?: string
-            }) =>
-              ({
-                ...e,
-                start: new Date(e.start),
-                end: new Date(e.end),
-              }) as ICalEvent
-          )
-          const { scheduleItems: merged, nextId } = icalEventsToScheduleItems(
-            events,
-            currentSchedule,
-            currentNextId,
-            WSU_SEMESTER.current.end,
-            url
-          )
-          currentSchedule = merged
-          currentNextId = nextId
-        }
-        setScheduleState((prev) => ({
-          ...prev,
-          scheduleItems: currentSchedule,
-          nextTaskId: currentNextId,
-        }))
-      } catch {
-        // Silent fail for background sync
-      }
-    }
-
-    const DAILY_MS = 24 * 60 * 60 * 1000
-
-    runBackgroundSync()
-
-    const intervalId = setInterval(runBackgroundSync, DAILY_MS)
-    return () => clearInterval(intervalId)
-  }, [icsUrls.length, setScheduleState])
+  useBackgroundIcsSync({ icsUrls, scheduleItems, nextTaskId, setScheduleState })
 
   if (showSurvey) {
     return (
